refactor(pdf-export): extract shared types and consolidate formatters with tests
